+++ MapAnalysis/RegionAnalysis/IRegion.cs
@@ -1,9 +1,9 @@
 ﻿using System.Numerics;
+using MapAnalysis.ExpandAnalysis;
 using SC2APIProtocol;
 
 namespace MapAnalysis.RegionAnalysis;
 
-// TODO GD This interface is rather "fat", but changing it represents a lot changes.
 public interface IRegion {
     /// <summary>
     /// The unique id of the region
@@ -27,13 +27,6 @@ public interface IRegion {
     public HashSet<Vector2> Cells { get; }
 
     /// <summary>
-    /// An approximation of the radius of the region.
-    /// TODO GD Get rid of this
-    /// </summary>
-    [Obsolete("Don't use ApproximatedRadius in new code")]
-    public float ApproximatedRadius { get; }
-
-    /// <summary>
     /// The type of this region.
     /// </summary>
     public RegionType Type { get; }
@@ -41,21 +34,10 @@ public interface IRegion {
     /// <summary>
     /// If the type of the region is RegionType.Expand, ExpandLocation will be the expand location of the region.
     /// </summary>
-    public IExpandLocation ExpandLocation { get; }
+    public IExpandLocation? ExpandLocation { get; }
 
     /// <summary>
-    /// Represents the neighbors of this region. A neighbor is describes as another Region and its frontier.
+    /// Represents the neighbors of this region. A neighbor is described as another Region and its frontier.
     /// </summary>
-    public IEnumerable<INeighboringRegion> Neighbors { get; }
-
-    /// <summary>
-    /// Whether the region is obstructed and the passage from one frontier to another is impossible.
-    /// </summary>
-    public bool IsObstructed { get; }
-
-    /// <summary>
-    /// Gets all the neighboring regions that can be reached.
-    /// </summary>
-    /// <returns>A list of regions that are accessible from this region.</returns>
-    public IEnumerable<IRegion> GetReachableNeighbors();
+    public IEnumerable<INeighboringRegion> Neighbors { get; } // TODO GD Change this to frontier
 }
